+++ app/build.gradle
@@ -43,9 +43,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.3.0'
     implementation "androidx.constraintlayout:constraintlayout:2.0.4"
     testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.mockito:mockito-core:3.11.0'
+    testImplementation 'org.mockito:mockito-inline:3.11.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-    testImplementation 'org.robolectric:robolectric:4.4'
-
 }
