+++ app/src/main/java/com/example/parkingproject/exceptions/ReleaseMoreThanOneMatchException.java
@@ -1,0 +1,8 @@
+package com.example.parkingproject.exceptions;
+
+public class ReleaseMoreThanOneMatchException extends Exception {
+
+    public ReleaseMoreThanOneMatchException() {
+        super("More than one match");
+    }
+}
