+++ app/src/main/java/com/example/parkingproject/exceptions/ReleaseNoMatchException.java
@@ -1,0 +1,7 @@
+package com.example.parkingproject.exceptions;
+
+public class ReleaseNoMatchException extends Exception {
+    public ReleaseNoMatchException() {
+        super("No matches for this reservation");
+    }
+}
