+++ app/src/main/java/com/example/parkingproject/exceptions/ReservationListEmptyException.java
@@ -1,0 +1,7 @@
+package com.example.parkingproject.exceptions;
+
+public class ReservationListEmptyException extends Exception {
+    public ReservationListEmptyException() {
+        super("Reservation list is empty");
+    }
+}
