+++ app/src/main/java/com/example/parkingproject/mvp/model/ParkingModel.java
@@ -1,0 +1,18 @@
+package com.example.parkingproject.mvp.model;
+
+public class ParkingModel {
+
+    private int parkingSize;
+
+    public int getParkingSize() {
+        return parkingSize;
+    }
+
+    public void setParkingSize(String newParkingSize) throws IllegalArgumentException {
+        int parseSize = Integer.parseInt(newParkingSize);
+        if (parseSize <= 0) {
+            throw new IllegalArgumentException();
+        }
+        this.parkingSize = parseSize;
+    }
+}
