+++ app/src/main/java/com/example/parkingproject/mvp/presenter/MenuPresenter.java
@@ -1,0 +1,20 @@
+package com.example.parkingproject.mvp.presenter;
+
+import com.example.parkingproject.entities.Parking;
+
+public class MenuPresenter {
+
+    private Parking parking;
+
+    public MenuPresenter(int parkingSize) {
+        parking = new Parking(parkingSize);
+    }
+
+    public Parking getParking() {
+        return parking;
+    }
+
+    public void setParking(Parking parking) {
+        this.parking = parking;
+    }
+}
