+++ app/src/main/java/com/example/parkingproject/mvp/presenter/ParkingPresenter.java
@@ -1,0 +1,28 @@
+package com.example.parkingproject.mvp.presenter;
+
+import com.example.parkingproject.mvp.model.ParkingModel;
+import com.example.parkingproject.mvp.view.ParkingView;
+
+public class ParkingPresenter {
+
+    private final ParkingModel parkingModel;
+    private final ParkingView parkingView;
+
+    public ParkingPresenter(ParkingModel parkingModel, ParkingView parkingView) {
+        this.parkingModel = parkingModel;
+        this.parkingView = parkingView;
+    }
+
+    public void onSizeCreationButtonPressed() {
+        String size = parkingView.getSize();
+        try {
+            parkingModel.setParkingSize(size);
+            int sizeInt = parkingModel.getParkingSize();
+            parkingView.navigateToMainMenuActivity(sizeInt);
+        } catch (NumberFormatException ex) {
+            parkingView.showErrorLargeNumber();
+        } catch (IllegalArgumentException ex) {
+            parkingView.showErrorLessEqual0();
+        }
+    }
+}
