+++ app/src/main/java/com/example/parkingproject/mvp/view/ActivityView.java
@@ -1,0 +1,33 @@
+package com.example.parkingproject.mvp.view;
+
+import android.app.Activity;
+import android.app.FragmentManager;
+import android.content.Context;
+
+import androidx.annotation.Nullable;
+
+import java.lang.ref.WeakReference;
+
+public class ActivityView {
+    private final WeakReference<Activity> activityRef;
+
+    public ActivityView(Activity activity) {
+        activityRef = new WeakReference<>(activity);
+    }
+
+    @Nullable
+    public Activity getActivity() {
+        return activityRef.get();
+    }
+
+    @Nullable
+    public Context getContext() {
+        return getActivity();
+    }
+
+    @Nullable
+    public FragmentManager getFragmentManager() {
+        Activity activity = getActivity();
+        return (activity != null) ? activity.getFragmentManager() : null;
+    }
+}
