--- app/src/main/java/com/example/parkingproyect/mvp/model/ParkingModel.java
@@ -1,18 +1,0 @@
-package com.example.parkingproyect.mvp.model;
-
-public class ParkingModel {
-
-    private int parkingSize;
-
-    public int getParkingSize() {
-        return parkingSize;
-    }
-
-    public void setParkingSize(String newParkingSize) throws IllegalArgumentException {
-        int parseSize = Integer.parseInt(newParkingSize);
-        if (parseSize <= 0) {
-            throw new IllegalArgumentException();
-        }
-        this.parkingSize = parseSize;
-    }
-}
