--- app/src/main/java/com/example/parkingproyect/mvp/view/ActivityView.java
@@ -1,34 +1,0 @@
-package com.example.parkingproyect.mvp.view;
-
-import android.app.Activity;
-import android.app.FragmentManager;
-import android.content.Context;
-
-import androidx.annotation.Nullable;
-
-
-import java.lang.ref.WeakReference;
-
-public class ActivityView {
-    private final WeakReference<Activity> activityRef;
-
-    public ActivityView(Activity activity) {
-        activityRef = new WeakReference<>(activity);
-    }
-
-    @Nullable
-    public Activity getActivity() {
-        return activityRef.get();
-    }
-
-    @Nullable
-    public Context getContext() {
-        return getActivity();
-    }
-
-    @Nullable
-    public FragmentManager getFragmentManager() {
-        Activity activity = getActivity();
-        return (activity != null) ? activity.getFragmentManager() : null;
-    }
-}
