--- app/src/main/java/com/example/parkingproyect/utils/Picker.java
@@ -1,62 +1,0 @@
-package com.example.parkingproyect.utils;
-
-import android.app.DatePickerDialog;
-import android.app.TimePickerDialog;
-import android.widget.DatePicker;
-import android.widget.EditText;
-import android.widget.TimePicker;
-
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
-
-public class Picker {
-
-    private DatePickerDialog datePickerDialog;
-    static final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yy-MM-dd HH:mm");
-    Date date;
-
-    public Date getDate() {
-        return date;
-    }
-
-    public Picker(EditText dateTime) {
-        showDateTimeDialog(dateTime);
-    }
-
-    public void showDateTimeDialog(EditText dateTime) {
-        Calendar calendar = Calendar.getInstance();
-        DatePickerDialog.OnDateSetListener dateSetListener = new DatePickerDialog.OnDateSetListener() {
-            @Override
-            public void onDateSet(DatePicker view, int year, int month, int dayOfMonth) {
-                calendar.set(Calendar.YEAR, year);
-                calendar.set(Calendar.MONTH, month);
-                calendar.set(Calendar.DAY_OF_MONTH, dayOfMonth);
-                TimePickerDialog.OnTimeSetListener timeSetListener = new TimePickerDialog.OnTimeSetListener() {
-                    @Override
-                    public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
-                        calendar.set(Calendar.HOUR_OF_DAY, hourOfDay);
-                        calendar.set(Calendar.MINUTE, minute);
-                        date = calendar.getTime();
-                        dateTime.setText(simpleDateFormat.format(calendar.getTime()));
-                    }
-                };
-                new TimePickerDialog(dateTime.getContext(),
-                        timeSetListener,
-                        calendar.get(Calendar.HOUR_OF_DAY),
-                        calendar.get(Calendar.MINUTE),
-                        false).show();
-            }
-        };
-        datePickerDialog = new DatePickerDialog(dateTime.getContext(),
-                dateSetListener,
-                calendar.get(Calendar.YEAR),
-                calendar.get(Calendar.MONTH),
-                calendar.get(Calendar.DAY_OF_MONTH));
-
-    }
-
-    public void show() {
-        datePickerDialog.show();
-    }
-}
